move util scripts out of commands folder
