+++ cli/lib/checks.py
